requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow==10.0.0
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
reportlab==4.1.0
